@@ -7,14 +7,15 @@ import '../../dtos/book_dto.dart';
 import '../remote_book_datasource_interfaces.dart';
 
 class LocalBookDataSource implements ILocalBookDataSource {
-  final ILocalDBProvider<Database> _localDBProvider;
   LocalBookDataSource({
     required ILocalDBProvider<Database> localDBProvider,
   }) : _localDBProvider = localDBProvider;
+  final ILocalDBProvider<Database> _localDBProvider;
 
   @override
   Future<BookEntity> favoriteBook({required BookEntity bookEntity}) async {
-    final book = BookDTO.fromEntity(bookEntity).copyWith(isFavorite: true);
+    final BookDTO book =
+        BookDTO.fromEntity(bookEntity).copyWith(isFavorite: true);
     await BookService.insertBook(
       bookDTO: book,
       database: await _localDBProvider.database,
@@ -23,17 +24,18 @@ class LocalBookDataSource implements ILocalBookDataSource {
   }
 
   @override
-  Future<List<BookEntity>> getBooksFavorite() async {
-    return await BookService.getAllBooks(
-      database: await _localDBProvider.database,
-    );
-  }
+  Future<List<BookEntity>> getBooksFavorite() async => BookService.getAllBooks(
+        database: await _localDBProvider.database,
+      );
 
   @override
   Future<BookEntity> unfavoriteBook({required BookEntity bookEntity}) async {
-    final book = BookDTO.fromEntity(bookEntity).copyWith(isFavorite: false);
+    final BookDTO book =
+        BookDTO.fromEntity(bookEntity).copyWith(isFavorite: false);
     await BookService.deleteBook(
-        bookId: bookEntity.id, database: await _localDBProvider.database);
+      bookId: bookEntity.id,
+      database: await _localDBProvider.database,
+    );
     return book;
   }
 }
