@@ -1,5 +1,4 @@
-import 'package:desafio_estante_de_livros/core/errors/exceptions.dart';
-
+import '../../../../../core/errors/exceptions.dart';
 import '../../../../../core/infrastructure/network/app_api.dart';
 import '../../../../../core/infrastructure/network/http_client/http_client_adapter_interface.dart';
 import '../../../domain/entities/book_entity.dart';
@@ -7,25 +6,33 @@ import '../../dtos/book_dto.dart';
 import '../remote_book_datasource_interfaces.dart';
 
 class RemoteBookDataSourceImpl implements IRemoteBookDataSource {
-  final IHttpClientAdapter _httpClientAdapter;
-
   RemoteBookDataSourceImpl({required IHttpClientAdapter httpClientAdapter})
       : _httpClientAdapter = httpClientAdapter;
+
+  final IHttpClientAdapter _httpClientAdapter;
+
   @override
-  Future<List<BookEntity>> getBooks() async {
-    final result = await _httpClientAdapter.get<List>(AppApi.getBooksURl);
+  Future<List<BookEntity>> getBooks({required int page}) async {
+    final Map<String, dynamic> result =
+        await _httpClientAdapter.get('${AppApi.getBooksURl}/?page=$page');
+    print(result);
     return List.generate(
-        result.length, (index) => BookDTO.fromMap(result[index]));
+      result['results'].length,
+      (index) => BookDTO.fromMap(result['results'][index]),
+    );
   }
 
   @override
-  Future<void> downloadBook(
-      {required String downloadUrl, required String pathDirectory}) async {
+  Future<void> downloadBook({
+    required String downloadUrl,
+    required String pathDirectory,
+  }) async {
     await _httpClientAdapter.download(
-        downloadUrl: downloadUrl,
-        pathDirectory: pathDirectory,
-        onError: () {
-          throw ServerException();
-        });
+      downloadUrl: downloadUrl,
+      pathDirectory: pathDirectory,
+      onError: () {
+        throw ServerException();
+      },
+    );
   }
 }
