@@ -7,7 +7,10 @@ abstract interface class ILocalBookDataSource {
 }
 
 abstract interface class IRemoteBookDataSource {
-  Future<List<BookEntity>> getBooks();
-  Future<void> downloadBook(
-      {required String downloadUrl, required String pathDirectory});
+  Future<List<BookEntity>> getBooks({required int page});
+
+  Future<void> downloadBook({
+    required String downloadUrl,
+    required String pathDirectory,
+  });
 }
