@@ -1,13 +1,6 @@
 import 'package:equatable/equatable.dart';
 
 class BookEntity extends Equatable {
-  final int id;
-  final String title;
-  final String author;
-  final String coverUrl;
-  final String downloadUrl;
-  final bool isFavorite;
-
   const BookEntity({
     required this.id,
     required this.title,
@@ -17,6 +10,16 @@ class BookEntity extends Equatable {
     required this.isFavorite,
   });
 
+  final int id;
+  final String title;
+  final String author;
+  final String coverUrl;
+  final String downloadUrl;
+  final bool isFavorite;
+
   @override
-  List<Object?> get props => [id];
+  List<Object?> get props => [
+        id,
+        isFavorite,
+      ];
 }
