@@ -3,11 +3,12 @@ import '../../repositories/book_repository_interface.dart';
 import 'favorite_book_usecase_interface.dart';
 
 class FavoriteBookUseCaseImpl implements IFavoriteBookUsecase {
-  final IBookRepository _bookRepository;
   FavoriteBookUseCaseImpl({required IBookRepository bookRepository})
       : _bookRepository = bookRepository;
+
+  final IBookRepository _bookRepository;
+
   @override
-  Future<BookEntity> call({required BookEntity bookEntity}) async {
-    return await _bookRepository.favoriteBook(bookEntity: bookEntity);
-  }
+  Future<BookEntity> call({required BookEntity bookEntity}) async =>
+      _bookRepository.favoriteBook(bookEntity: bookEntity);
 }
