@@ -1,5 +1,5 @@
 import '../../entities/book_entity.dart';
 
 abstract interface class IGetBooksUsecase {
-  Future<List<BookEntity>> call();
+  Future<List<BookEntity>> call({int page});
 }
