@@ -1,9 +1,9 @@
 import 'dart:async';
 
 import 'package:bloc/bloc.dart';
-import 'package:desafio_estante_de_livros/core/errors/messages.dart';
 import 'package:equatable/equatable.dart';
 
+import '../../../../../../core/errors/messages.dart';
 import '../../../../data/dtos/book_dto.dart';
 import '../../../../domain/entities/book_entity.dart';
 import '../../../../domain/use_cases/favorite_book/favorite_book_usecase_interface.dart';
@@ -14,11 +14,6 @@ import '../../../../domain/use_cases/un_favorite_book/un_favorite_book_usecase_i
 part 'home_state.dart';
 
 class HomeController extends Cubit<HomeState> {
-  final IGetBooksUsecase _getBooksUsecase;
-  final IFavoriteBookUsecase _favoriteBookUsecase;
-  final IUnFavoriteBookUsecase _unFavoriteBookUsecase;
-  final IGetFavoritesBooksUsecase _getFavoritesBooksUsecase;
-
   HomeController({
     required IGetBooksUsecase getBooksUsecase,
     required IFavoriteBookUsecase favoriteBookUsecase,
@@ -30,40 +25,66 @@ class HomeController extends Cubit<HomeState> {
         _getFavoritesBooksUsecase = getFavoritesBooksUsecase,
         super(const HomeState.initial());
 
-  Future<void> fetchAllBooks() async {
-    emit(state.copyWith(status: HomeStatus.loading));
-    final favorites = await _fetchFavoritesBooks();
-    final booksApi = await _fecthBooksApi();
-    final result = booksApi
-        .map((e) {
-          if (favorites.contains(e)) {
-            return BookDTO.fromEntity(e).copyWith(isFavorite: true);
-          }
-          return e;
-        })
-        .toSet()
-        .toList();
-    emit(state.copyWith(books: result, status: HomeStatus.loaded));
-  }
+  final IGetBooksUsecase _getBooksUsecase;
+  final IFavoriteBookUsecase _favoriteBookUsecase;
+  final IUnFavoriteBookUsecase _unFavoriteBookUsecase;
+  final IGetFavoritesBooksUsecase _getFavoritesBooksUsecase;
 
-  Future<List<BookEntity>> _fecthBooksApi() async {
+  Future<List<BookEntity>> _fecthBooksApi({int page = 1}) async {
     try {
-      return await _getBooksUsecase.call();
+      return await _getBooksUsecase.call(page: page);
     } catch (e) {
-      emit(state.copyWith(
+      emit(
+        state.copyWith(
           status: HomeStatus.failure,
-          errorMessage: ErrorMessages.getBooksError));
+          errorMessage: ErrorMessages.getBooksError,
+        ),
+      );
       return List.empty();
     }
+  }
+
+  Future<void> fetchNextPage({bool reset = false}) async {
+    if (state.isLoadingMore || !state.hasMore) return;
+
+    emit(state.copyWith(status: HomeStatus.loadingMore));
+
+    final int nextPage = state.page + 1;
+    final List<BookEntity> favorites = await _fetchFavoritesBooks();
+    final List<BookEntity> newBooks = await _fecthBooksApi(page: nextPage);
+
+    if (newBooks.isEmpty) {
+      emit(state.copyWith(status: HomeStatus.loaded, hasMore: false));
+      return;
+    }
+
+    final List<BookEntity> mergedBooks = reset ? [] : [...state.books];
+    for (final book in newBooks) {
+      final bool isFav = favorites.any((fav) => fav.id == book.id);
+      final BookEntity bookWithFav =
+          isFav ? BookDTO.fromEntity(book).copyWith(isFavorite: true) : book;
+      mergedBooks.add(bookWithFav);
+    }
+
+    emit(
+      state.copyWith(
+        books: mergedBooks.toSet().toList(),
+        status: HomeStatus.loaded,
+        page: nextPage,
+      ),
+    );
   }
 
   Future<List<BookEntity>> _fetchFavoritesBooks() async {
     try {
       return await _getFavoritesBooksUsecase.call();
     } catch (e) {
-      emit(state.copyWith(
+      emit(
+        state.copyWith(
           status: HomeStatus.failure,
-          errorMessage: ErrorMessages.getFavoritesBooksError));
+          errorMessage: ErrorMessages.getFavoritesBooksError,
+        ),
+      );
       return List.empty();
     }
   }
@@ -78,33 +99,60 @@ class HomeController extends Cubit<HomeState> {
 
   Future<void> favoriteBook({required BookEntity bookEntity}) async {
     try {
-      final book = await _favoriteBookUsecase.call(bookEntity: bookEntity);
-      final result = state.books
-          .map((e) => e == book
-              ? BookDTO.fromEntity(e).copyWith(isFavorite: book.isFavorite)
-              : e)
+      final BookEntity book =
+          await _favoriteBookUsecase.call(bookEntity: bookEntity);
+      final List<BookEntity> result = state.books
+          .map(
+            (e) => e.id == book.id
+                ? BookDTO.fromEntity(e).copyWith(isFavorite: book.isFavorite)
+                : e,
+          )
           .toList();
       emit(state.copyWith(books: result, status: HomeStatus.loaded));
     } catch (e) {
-      emit(state.copyWith(
+      emit(
+        state.copyWith(
           status: HomeStatus.failure,
-          errorMessage: ErrorMessages.favoriteBookError));
+          errorMessage: ErrorMessages.favoriteBookError,
+        ),
+      );
     }
   }
 
   Future<void> _unFavoriteBook({required BookEntity bookEntity}) async {
     try {
-      final book = await _unFavoriteBookUsecase.call(bookEntity: bookEntity);
-      final result = state.books
-          .map((e) => e == book
-              ? BookDTO.fromEntity(e).copyWith(isFavorite: book.isFavorite)
-              : e)
+      final BookEntity book =
+          await _unFavoriteBookUsecase.call(bookEntity: bookEntity);
+      final List<BookEntity> result = state.books
+          .map(
+            (e) => e.id == book.id
+                ? BookDTO.fromEntity(e).copyWith(isFavorite: book.isFavorite)
+                : e,
+          )
           .toList();
       emit(state.copyWith(books: result, status: HomeStatus.loaded));
     } catch (e) {
-      emit(state.copyWith(
+      emit(
+        state.copyWith(
           status: HomeStatus.failure,
-          errorMessage: ErrorMessages.unfavoreteBookError));
+          errorMessage: ErrorMessages.unfavoreteBookError,
+        ),
+      );
     }
+  }
+
+  Future<void> reset() async {
+    emit(
+      state.copyWith(
+        status: HomeStatus.initial,
+        page: 0,
+        hasMore: true,
+      ),
+    );
+  }
+
+  Future<void> refresh() async {
+    await reset();
+    await fetchNextPage(reset: true);
   }
 }
