@@ -1,18 +1,13 @@
 part of 'home_controller.dart';
 
-enum HomeStatus { initial, loading, loaded, failure }
+enum HomeStatus { initial, loading, loadingMore, loaded, failure }
 
 class HomeState extends Equatable {
-  final HomeStatus status;
-  final List<BookEntity> books;
-  final String? errorMessage;
-
-  List<BookEntity> get favoritesBooks =>
-      books.where((element) => element.isFavorite).toList();
-
   const HomeState._({
     required this.status,
     required this.books,
+    required this.page,
+    required this.hasMore,
     this.errorMessage,
   });
 
@@ -20,20 +15,36 @@ class HomeState extends Equatable {
       : this._(
           books: const [],
           status: HomeStatus.initial,
+          page: 0,
+          hasMore: true,
         );
 
+  final HomeStatus status;
+  final List<BookEntity> books;
+  final String? errorMessage;
+  final int page;
+  final bool hasMore;
+
+  List<BookEntity> get favoritesBooks =>
+      books.where((element) => element.isFavorite).toList();
+
+  bool get isLoadingMore => status == HomeStatus.loadingMore;
+
   @override
-  List<Object?> get props => [status, errorMessage, books, favoritesBooks];
+  List<Object?> get props => [status, books, errorMessage, page, hasMore];
 
   HomeState copyWith({
     HomeStatus? status,
     List<BookEntity>? books,
     String? errorMessage,
-  }) {
-    return HomeState._(
-      status: status ?? this.status,
-      books: books ?? this.books,
-      errorMessage: errorMessage,
-    );
-  }
+    int? page,
+    bool? hasMore,
+  }) =>
+      HomeState._(
+        books: books ?? this.books,
+        status: status ?? this.status,
+        errorMessage: errorMessage,
+        page: page ?? this.page,
+        hasMore: hasMore ?? this.hasMore,
+      );
 }
