@@ -8,14 +8,16 @@ import '../controllers/home_controller.dart';
 import 'widgets/books_tab_view.dart';
 
 class HomePage extends StatefulWidget {
-  const HomePage(
-      {super.key,
-      required HomeController homeController,
-      required DownloadBookController downloadBookController})
-      : _homeController = homeController,
+  const HomePage({
+    required HomeController homeController,
+    required DownloadBookController downloadBookController,
+    super.key,
+  })  : _homeController = homeController,
         _downloadBookController = downloadBookController;
+
   final HomeController _homeController;
   final DownloadBookController _downloadBookController;
+
   @override
   State<HomePage> createState() => _HomePageState();
 }
@@ -23,13 +25,13 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   @override
   void initState() {
-    widget._homeController.fetchAllBooks();
+    widget._homeController.fetchNextPage();
     super.initState();
   }
 
   void listenState(BuildContext context, HomeState state) {
     if (state.errorMessage != null) {
-      AsukaSnackbar.warning(state.errorMessage ?? "").show();
+      AsukaSnackbar.warning(state.errorMessage ?? '').show();
     }
   }
 
@@ -38,31 +40,33 @@ class _HomePageState extends State<HomePage> {
     return MultiBlocProvider(
       providers: [
         BlocProvider(create: (context) => widget._homeController),
-        BlocProvider(create: (context) => widget._downloadBookController)
+        BlocProvider(create: (context) => widget._downloadBookController),
       ],
       child: DefaultTabController(
         length: 2,
         child: Scaffold(
           appBar: AppBar(
-            title: const AppTitleText(title: "Minha Estante "),
-            bottom: TabBar(tabs: [
-              const Tab(text: "Livros"),
-              BlocBuilder<HomeController, HomeState>(
-                builder: (context, state) {
-                  return Badge.count(
-                      isLabelVisible: state.favoritesBooks.isNotEmpty,
-                      largeSize: 15,
-                      count: state.favoritesBooks.length,
-                      child: const Tab(text: "Favoritos"));
-                },
-              ),
-            ]),
+            title: const AppTitleText(title: 'Minha Estante '),
+            bottom: TabBar(
+              tabs: [
+                const Tab(text: 'Livros'),
+                BlocConsumer<HomeController, HomeState>(
+                  listener: (context, state) {
+                    print('Listener called with state: $state');
+                  },
+                  builder: (context, state) => Badge.count(
+                    isLabelVisible: state.favoritesBooks.isNotEmpty,
+                    largeSize: 15,
+                    count: state.favoritesBooks.length,
+                    child: const Tab(text: 'Favoritos'),
+                  ),
+                ),
+              ],
+            ),
           ),
           body: SizedBox(
             height: MediaQuery.sizeOf(context).height,
             child: BlocConsumer<HomeController, HomeState>(
-              listenWhen: (previous, current) =>
-                  previous.status != current.status,
               listener: listenState,
               builder: (context, state) {
                 if (state.status == HomeStatus.loading) {
