@@ -1,15 +1,12 @@
 abstract class ErrorMessages {
-  static const String downloadBookError =
-      "Não foi possível fazer o download do livro";
-  static const String androidVersionError =
-      "Não foi possível pegar a versão do android";
-  static const String directoryDenied = "O acesso do diretório foi negado";
-  static const String platformNotFound =
-      "Ocorreu um erro ao identificar a plataforma";
-  static const String getBooksError = "Não foi possível resgatar os livros";
-  static const String getFavoritesBooksError =
-      "Não foi possível resgatar os livros favoritos";
-  static const String favoriteBookError = "Não foi possível favoritar o livro";
-  static const String unfavoreteBookError =
-      "Não foi possível desfavoritar o livro";
+  static const downloadBookError = 'Não foi possível fazer o download do livro';
+  static const androidVersionError =
+      'Não foi possível pegar a versão do android';
+  static const directoryDenied = 'O acesso do diretório foi negado';
+  static const platformNotFound = 'Ocorreu um erro ao identificar a plataforma';
+  static const getBooksError = 'Não foi possível resgatar os livros';
+  static const getFavoritesBooksError =
+      'Não foi possível resgatar os livros favoritos';
+  static const favoriteBookError = 'Não foi possível favoritar o livro';
+  static const unfavoreteBookError = 'Não foi possível desfavoritar o livro';
 }
