@@ -1,16 +1,16 @@
-import '../../theme/app_colors.dart';
 import 'package:vocsy_epub_viewer/epub_viewer.dart';
+
+import '../../theme/app_colors.dart';
 
 class EpubViewService {
   EpubViewService._();
 
   static EpubViewService get instance => EpubViewService._();
 
-  config() {
+  void config() {
     VocsyEpub.setConfig(
       themeColor: AppColors.primaryColor,
-      identifier: "iosBook",
-      scrollDirection: EpubScrollDirection.ALLDIRECTIONS,
+      identifier: 'iosBook',
       allowSharing: true,
       enableTts: true,
       nightMode: true,
@@ -22,10 +22,10 @@ class EpubViewService {
     VocsyEpub.open(
       path,
       lastLocation: EpubLocator.fromJson({
-        "bookId": "2239",
-        "href": "/OEBPS/ch06.xhtml",
-        "created": 1539934158390,
-        "locations": {"cfi": "epubcfi(/0!/4/4[simple_book]/2/2/6)"}
+        'bookId': '2239',
+        'href': '/OEBPS/ch06.xhtml',
+        'created': 1539934158390,
+        'locations': {'cfi': 'epubcfi(/0!/4/4[simple_book]/2/2/6)'},
       }),
     );
   }
