@@ -1,4 +1,4 @@
 abstract class Routes {
-  static const String initialRoute = "/";
-  static const String bookDeatils = "book-details";
+  static const initialRoute = '/';
+  static const bookDeatils = 'book-details';
 }
