@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 
 abstract class AppColors {
-  static const Color primaryColor = Color(0xFF1B9DFF);
-  static const Color dangerColor = Color(0xFFFF4455);
-  static const Color red = Color(0xFFFF4455);
-  static const Color blackColor = Color(0xFF000000);
-  static const Color whiteColor = Color(0xFFFFFFFF);
-  static const Color labelColor = Color(0xFFA9A7A7);
-  static const Color borderColor = Color(0xFFA9A7A7);
-  static const Color bookTitleColor = Color(0xFF565656);
-  static const Color titleColor = Color(0xFF575757);
-  static const Color transparent = Color(0x00000000);
+  static const primaryColor = Color(0xFF1B9DFF);
+  static const dangerColor = Color(0xFFFF4455);
+  static const red = Color(0xFFFF4455);
+  static const blackColor = Color(0xFF000000);
+  static const whiteColor = Color(0xFFFFFFFF);
+  static const labelColor = Color(0xFFA9A7A7);
+  static const borderColor = Color(0xFFA9A7A7);
+  static const bookTitleColor = Color(0xFF565656);
+  static const titleColor = Color(0xFF575757);
+  static const transparent = Color(0x00000000);
 }
