@@ -1,56 +1,56 @@
 abstract class AppSizes {
-  static const double size00 = 0.0;
-  static const double size05 = 5.0;
-  static const double size10 = 10.0;
-  static const double size15 = 15.0;
-  static const double size20 = 20.0;
-  static const double size25 = 25.0;
-  static const double size30 = 30.0;
-  static const double size35 = 35.0;
-  static const double size40 = 40.0;
-  static const double size45 = 45.0;
-  static const double size50 = 50.0;
-  static const double size55 = 55.0;
-  static const double size60 = 60.0;
-  static const double size65 = 65.0;
-  static const double size70 = 70.0;
-  static const double size80 = 80.0;
-  static const double size90 = 90.0;
-  static const double size100 = 100.0;
-  static const double size120 = 120.0;
-  static const double size140 = 140.0;
-  static const double size150 = 150.0;
-  static const double size200 = 200.0;
-  static const double size210 = 210.0;
-  static const double size230 = 230.0;
-  static const double size240 = 240.0;
-  static const double size250 = 250.0;
-  static const double size285 = 285.0;
-  static const double size300 = 300.0;
+  static const double size00 = 0;
+  static const double size05 = 5;
+  static const double size10 = 10;
+  static const double size15 = 15;
+  static const double size20 = 20;
+  static const double size25 = 25;
+  static const double size30 = 30;
+  static const double size35 = 35;
+  static const double size40 = 40;
+  static const double size45 = 45;
+  static const double size50 = 50;
+  static const double size55 = 55;
+  static const double size60 = 60;
+  static const double size65 = 65;
+  static const double size70 = 70;
+  static const double size80 = 80;
+  static const double size90 = 90;
+  static const double size100 = 100;
+  static const double size120 = 120;
+  static const double size140 = 140;
+  static const double size150 = 150;
+  static const double size200 = 200;
+  static const double size210 = 210;
+  static const double size230 = 230;
+  static const double size240 = 240;
+  static const double size250 = 250;
+  static const double size285 = 285;
+  static const double size300 = 300;
 }
 
 abstract class BorderSize {
-  static const double small = 1.0;
-  static const double medium = 2.0;
-  static const double large = 3.0;
+  static const double small = 1;
+  static const double medium = 2;
+  static const double large = 3;
 }
 
 abstract class BorderRadiusSize {
-  static const double small = 12.0;
-  static const double medium = 15.0;
-  static const double large = 20.0;
+  static const double small = 12;
+  static const double medium = 15;
+  static const double large = 20;
 }
 
 abstract class FontSize {
-  static const double fontSize12 = 12.0;
-  static const double fontSize14 = 14.0;
-  static const double fontSize16 = 16.0;
-  static const double fontSize20 = 20.0;
-  static const double fontSize22 = 22.0;
-  static const double fontSize24 = 24.0;
-  static const double fontSize28 = 28.0;
-  static const double fontSize32 = 32.0;
-  static const double fontSize36 = 36.0;
-  static const double fontSize45 = 45.0;
-  static const double fontSize57 = 57.0;
+  static const double fontSize12 = 12;
+  static const double fontSize14 = 14;
+  static const double fontSize16 = 16;
+  static const double fontSize20 = 20;
+  static const double fontSize22 = 22;
+  static const double fontSize24 = 24;
+  static const double fontSize28 = 28;
+  static const double fontSize32 = 32;
+  static const double fontSize36 = 36;
+  static const double fontSize45 = 45;
+  static const double fontSize57 = 57;
 }
